Removed extra TF code

+++ shared_functions.py
@@ -8,11 +8,6 @@ from typing import List
 
 from data_transformations import normalize_2D_array_inplace, standardize_2D_array_inplace, map_2D_array
 
-os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
-import tensorflow as tf
-import tensorflow_addons as tfa
-import tensorflow_probability as tfp
-
 
 def load_mat(in_path: str, standardize=False, normalize=False, normalize_range=(-1, 1), channels=[-1], use_cached=True, debug_print=False) -> np.ndarray:
     restricted_chars = "/!@#$%^&*()=+[]{}\'\";:,.<>`~ "
@@ -34,20 +29,25 @@ def load_mat(in_path: str, standardize=False, normalize=False, normalize_range=(
         print(f"One-time processing {processed_path}...")
         # if we need to create the file, save it at the end
         try:
+
+            # Runs on raw files
             # loaded_mat = scipy.io.loadmat(in_path)
             # Find the data's name of the Matlab dictionary,
             # assuming it's the last key...
             # tag = list(loaded_mat)[-1]
             # mat = np.array(loaded_mat[tag])
+
+            # Runs on non-raw files
             mat = np.array(h5py.File(in_path)['EEG']['data'])
-            mat=np.squeeze(mat)
+            mat = np.squeeze(mat)
 
 
         except Exception as e:
             print(f"Exception! Could not loadmat {in_path}. e: {e}")
-            return None
-            # mat = [-1 for _ in range(len(channels))]
+            mat = np.array([-1 for _ in range(len(channels))])
+            return mat
 
+        # for non-raw files
         # Tries to recursively find the electrode readings by finding the deepest and largest subarray
         # while True:
         #     try:
@@ -84,13 +84,15 @@ def load_mat(in_path: str, standardize=False, normalize=False, normalize_range=(
                 print()
 
         mat = np.array(mat)
-        # Saves the processed file on the disk
-        f = open(processed_path, 'w')
-        try:
-            np.save(file=processed_path, arr=mat, )
-        except Exception as e:
-            print(f"Exception!: Failed to save {processed_path}. e: {e}")
-            os.remove(processed_path)
+
+        if use_cached:
+            # Saves the processed file on the disk
+            f = open(processed_path, 'w')
+            try:
+                np.save(file=processed_path, arr=mat, )
+            except Exception as e:
+                print(f"Exception!: Failed to save {processed_path}. e: {e}")
+                os.remove(processed_path)
 
         return mat
 
@@ -111,37 +113,3 @@ def find_files(in_path):
                 files.append(each)
 
     return files
-
-
-class ActivationLayerFactory():
-    def __init__(self, activation=tf.keras.activations.relu):
-        self.__activation_layer_counter = 0
-        self.activation = activation
-
-    def produce(self, activation=None):
-        if activation is None: activation = tf.keras.activations.relu
-        temp = tf.keras.layers.Activation(activation=activation, name='activation_' + str(self.__activation_layer_counter))
-        self.__activation_layer_counter += 1
-        return temp
-
-
-class DropoutLayerFactory():
-    def __init__(self, dropout_rate=0.5):
-        self.__dropout_layer_counter = 0
-        self.rate = dropout_rate
-
-    def produce(self, rate=None):
-        if rate == None: rate = self.rate
-        temp = tf.keras.layers.Dropout(rate=rate, name='dropout' + str(self.__dropout_layer_counter))
-        self.__dropout_layer_counter += 1
-        return temp
-
-
-class BatchNormalizationLayerFactory():
-    def __init__(self):
-        self.__batchnormalization_layer_count = 0
-
-    def produce(self):
-        temp = tf.keras.layers.BatchNormalization(name='BN' + str(self.__batchnormalization_layer_count))
-        self.__batchnormalization_layer_count += 1
-        return temp
